requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cloudinary==1.41.0
dj-database-url==2.2.0
Django==5.1
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
pytest==8.3.2
pytest-django==4.8.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
